test: Fix test suite bugs for 16-bit targets
Use uintptr_t for pointers converted to scalar, use int32_t for
32-bit values.

+++ newlib/testsuite/newlib.time/tzset.c
@@ -7,13 +7,13 @@
 #include <time.h>
 #include <limits.h>
 
-#define IN_SECONDS(h, m, s) ((h) * 3600 + (m) * 60 + (s))
+#define IN_SECONDS(h, m, s) ((int32_t) (h) * 3600 + (m) * 60 + (s))
 #define NO_TIME INT_MIN
 
 struct tz_test {
     const char* tzstr;
-    int offset_seconds;
-    int dst_offset_seconds;
+    int32_t offset_seconds;
+    int32_t dst_offset_seconds;
 };
 
 extern struct tm winter_tm;
